// dmtApp/src/app/dmt/interfaces/empleado.interface.ts
export interface Empleado {
    NUMERO?:    number;
    NOMBRE:    string;
    RFC:       string;
    DIRECCION?: string;
    CIUDAD?:    string;
    ESTADO?:    string;
    NACIOEN?:   string;
    CIVIL?:     string;
    ESTATURA?:  number;
    PESO?:      number;
    SEXO?:      string;
    TELEFONO?:  string;
    AVISAR?:    string;
    PUESTO?:    string;
    DEPARTA?:   string;
    IMSS?:      string;
    TIPSANGRE?: string;
    EMPRESA?:   number;
    NOMINA?:    number;
    FECING?:    Date;
    CARTILLA?:  string;
    LICENCIA?:  string;
    PASAPORTE?: string;
    LI_INICIO?: Date;
    LI_VENCE?:  Date;
    PA_VENCE?:  Date;
    ADUANA?:    string;
    AD_VENCE?:  Date;
    CURP?:      string;
    CR_NORTE?:  string;
    CR_SUR?:    string;
    CR_ESTE?:   string;
    CR_OESTE?:  string;
    FEC_NACIO?: Date;
    FEC_BAJA?:  Date;
    COLONIA?:   string;
    CODPOS?:    string;
    operador?:  string;
    ufolio?:    number;
    tel_fam?:   string;
    tel_tbj?:   string;
    observa?:   string;
    correo?:    string;
}